+++ inference_perf/metrics/__init__.py
@@ -11,7 +11,7 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-from .base import MetricsSource
+from .base import MetricsSource, RequestMetric
 from .observed import ObservedMetricsCollector
 
-__all__ = ["MetricsSource", "ObservedMetricsCollector"]
+__all__ = ["MetricsSource", "ObservedMetricsCollector", "RequestMetric"]
